The reading of sensor is converted to a function (non-blocking) which returns false if the valid bit is not set else returns true..The MCU is put to sleep at the end of every accquisition and woken up at timer interrupt i.e. as per the accquisition interval set by the user

+++ au20xx_uart_fixedpoint/au20xx_flow_sensor/source/common.c
@@ -463,7 +463,7 @@ uint32_t average_by_4(uint32_t * address_of_array, int* slope_of_slope)
 * <hr>
 *
 *******************************************************************************/
-int average_by_8(int * address_of_array)
+int8_t average_by_8(int8_t * address_of_array)
 {
     static uint8_t i = 0;
     static int average_value;
